Add files via upload
The   "Online Food Ordering   System" has been   developed to override   the  problemsprevailing in the practicing manual system. This software is supported to eliminate and,in  some cases, reduce  the  hardships  faced   by this  existing  system.  Moreover,   thissystem is designed for the particular need of the company to carry out operations in asmooth and effective manner